#726: Remove NFTokenState and update NameState class
