+++ core/src/main/java/io/neow3j/protocol/core/response/NameState.java
@@ -1,24 +1,49 @@
 package io.neow3j.protocol.core.response;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
+import io.neow3j.types.Hash160;
 
 import java.util.Objects;
 
-public class NameState extends NFTokenState {
+public class NameState {
+
+    @JsonProperty(value = "name", required = true)
+    private String name;
 
     @JsonProperty(value = "expiration")
     private Long expiration;
 
+    @JsonProperty(value = "admin")
+    private Hash160 admin;
+
     public NameState() {
     }
 
-    public NameState(String name, Long expiration) {
-        super(name);
+    public NameState(String name, Long expiration, Hash160 admin) {
+        this.name = name;
         this.expiration = expiration;
+        this.admin = admin;
     }
 
+    /**
+     * @return the name.
+     */
+    public String getName() {
+        return name;
+    }
+
+    /**
+     * @return the expiration.
+     */
     public Long getExpiration() {
         return expiration;
+    }
+
+    /**
+     * @return the admin.
+     */
+    public Hash160 getAdmin() {
+        return admin;
     }
 
     @Override
@@ -31,12 +56,13 @@ public class NameState extends NFTokenState {
         }
         NameState that = (NameState) o;
         return Objects.equals(getName(), that.getName()) &&
-                Objects.equals(getExpiration(), that.getExpiration());
+                Objects.equals(getExpiration(), that.getExpiration()) &&
+                Objects.equals(getAdmin(), that.getAdmin());
     }
 
     @Override
     public int hashCode() {
-        return Objects.hash(getName(), getExpiration());
+        return Objects.hash(getName(), getExpiration(), getAdmin());
     }
 
     @Override
@@ -44,6 +70,7 @@ public class NameState extends NFTokenState {
         return "Properties{" +
                 "name='" + getName() + '\'' +
                 ", expiration=" + expiration +
+                ", admin='" + getAdmin().toAddress() +
                 "}";
     }
